fix: ui change
small changes to the ui
